sns-management/requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
tweepy==4.14.0
schedule==1.2.0
python-dotenv==1.0.0